requirements.txt
google-cloud-secret-manager==2.16.2
google-cloud-logging==3.5.0
google-cloud-monitoring==2.15.1
google-cloud-bigquery==3.17.0
requests==2.31.0
fastapi==0.104.0
uvicorn==0.23.2
pytz==2023.3.post1
pydantic
starlette==0.27.0
# Test/dev dependencies
pytest==8.4.2
python-dotenv==1.0.0
pytest-asyncio==1.2.0
